Improve CPU latency QoS handling
Increase the default latency QoS request from 0 to 200us, and add a
module parameter to tune it if necessary.

Try to predict when the next DMA transfer will occur, and only enable
QoS just before that, and only if we've actually seen any DMA errors.

Warn when a large number of DMA transfer errors occur.